image_recognition_server/requirements.txt
flask==3.0.0
flask-cors==4.0.0
transformers==4.36.0
torch==2.1.2
pillow>=10.0.0
gunicorn==23.0.0
